+++ src/AdminMenu/TranslatePageEventPropagationBuilder.php
@@ -31,12 +31,7 @@ class TranslatePageEventPropagationBuilder implements PropagationInterface
         $this->eventDispatcher = $eventDispatcher;
     }
 
-    /**
-     * Build the relevant event and forward it.
-     *
-     * @return mixed|void
-     */
-    public function buildAndForwardEvent(Event $event)
+    public function buildAndForwardEvent(Event $event): void
     {
         if (!$event instanceof PageViewEvent) {
             return;