+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added|Changed|Deprecated|Removed|Fixed|Security
 Nothing so far
 
+## 7.0.0 - 2022-07-11
+### Removed
+- Support for Sonata ^3
+- Removed compatibility with Zicht Versioning bundle
+### Added
+- Support for Sonata ^4
+- Added more types (e.g. return types)
+
 ## 6.4.2 - 2022-10-04
 ### Changed
 - Introduced PHP CS Fixer + fixed all code in src/ and tests/.